Set buffer (and chunk) size to 8 KiB when streaming files (#152)
The previous implementation uses `BytesMut::new` to create a value with
"length 0 and unspecified capacity".  The documentation for
`BytesMut::with_capacity` states that "If
`capacity` is under `4 * size_of::<usize>() - 1`, then BytesMut will
not allocate.

By allocating here, each chunk that is processed will allocate a new
buffer.  The buffer size of 8 KiB was chosen arbitrarily, but attempts
to weight per-request overhead against the number of allocations
needed to serve a file.